requirements.txt
paquo==0.4.0
openslide-python==1.1.2
Pillow==8.4.0
dynaconf==3.1.7
pytest-cov==3.0.0
numpy==1.21.4
psutil==5.8.0
pyvips==2.1.16
line-profiler==3.3.1
memory-profiler==0.58.0
opencv-python==4.5.4.60